requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.1.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.4
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.1
uritemplate==4.2.0
Werkzeug==3.1.3
